Source/requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
fastapi==0.112.1
uvicorn==0.30.6
aiohttp~=3.10.5
Pillow==10.4.0
aiomysql==0.2.0
replicate==0.32.0
